Add files via upload
I cleaned up the code from the previous crawler.c file and added comments to explain what each part does.
I did not implement tidy yet but I think it would be a good library to use for parsing html.
We need to implement multithreading. Right now the program should only use one thread.
We need to implement an output file. Right now the program does not output "anything".
So we need to implement it where it outputs the url and the html to a file.

So we can use the other file as a reference for multithreading. Since it was messy. 
We know what parts of the code we need to use for multithreading. So we can just add those parts into this file.

What we need to my knowledge is:
1. A multithread implementation of the crawler function
2. A multithread implementation of the queue functions
3. A output function that outputs the url and html to a file
4. A parser function that parses the html and stores the urls in the queue
---This is where tidy comes in. We can use tidy to parse the html and store the urls in the queue
5. Add to main function that initializes the queue and threads and calls the crawler function