crawler using tidy library
Initially we wanted to include the tidy library in our project but, we decided to scrap the idea.  Doing some research on Tidy I attempted to use it in this version of a web crawler in c. Here are the resources I used to help me create this crawler using the library tidy: 

 https://www.html-tidy.org/developer/