Create curl_test1.c
All I did was add the pthread library in an attempt to get the multithread functional. I did add some changes within the main function as well (which I state in the comments in the vode). Also added an additional function fetch_page to help with fetching pages via curl and adding more link to the multihandle.